% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisCon_bivariatesmooth.R
\name{BasisCon.bivariate.smooth}
\alias{BasisCon.bivariate.smooth}
\title{Basis Construction Function}
\usage{
\method{BasisCon}{bivariate.smooth}(object, data)
}
\arguments{
\item{object}{is a bivariate smooth object.}

\item{data}{A data frame, model frame or list containing the values of the (named) covariates
at which the smooth term is to be evaluated. If it’s a list then \code{n} must be supplied.}
}
\value{
a list of \code{smooth} objects are returned. Many of the outputs are from \code{f} function.
Other outputs include all the information related to berstein basis.
}
\description{
This is an internal function of package \code{ggam}.
}
\details{
It is the wrapper function which calls basis constructing method.
}
\examples{
library(BPST)
library(GgAM)
data("eg1pop_dat")
eg1_V2=eg1pop_dat[['V2']]
eg1_T2=eg1pop_dat[['T2']]
eg1pop_rho03=eg1pop_dat[['rho03']]
sam=eg1pop_rho03[sample(1:dim(eg1pop_rho03)[1],70),]
BI <- BasisCon(b(x1,x2,d=2,r=1,V=eg1_V2,Tr=eg1_T2),eg1pop_rho03[sample(1:dim(eg1pop_rho03)[1],70),])
xx=seq(-0.8,3.8,length=100)
yy=seq(-0.8,0.8,length=100)
sam=data.frame(x1=rep(xx,length(xx)),x2=rep(yy,each=length(yy)))
BI <- BasisCon(b(x1,x2,d=2,r=1,V=eg1_V2,Tr=eg1_T2),sam)
}
