% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisCon_univariatesmooth.R
\name{BasisCon.univariate.smooth}
\alias{BasisCon.univariate.smooth}
\title{Basis Construction Function}
\usage{
\method{BasisCon}{univariate.smooth}(object, data)
}
\arguments{
\item{object}{is a smooth specification object or a smooth object.}

\item{data}{A data frame, model frame or list containing the values of the (named) covariates
at which the smooth term is to be evaluated. If it’s a list then \code{n} must be supplied.}
}
\value{
a list of \code{smooth} objects are returned. Many of the outputs are from \code{f} function.
Other outputs include all the information related to berstein basis.
}
\description{
This is an internal function of package \code{ggam}.
}
\details{
It is the wrapper function which calls basis constructing method.
}
\examples{
library(BPST)
library(GgAM)
data("eg1pop_dat")
eg1_V2=eg1pop_dat[['V2']]
eg1_T2=eg1pop_dat[['T2']]
eg1pop_rho03=eg1pop_dat[['rho03']]
sam=eg1pop_rho03[sample(1:dim(eg1pop_rho03)[1],70),]
aa=BasisCon(u(z1),sam)
}
