% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_se.R
\name{beta_se}
\alias{beta_se}
\title{Standard Error Estimation for Linear Coefficients}
\usage{
beta_se(UB, B, Q2, K, lam1, lam2, Z, Y, ind.c, VS = FALSE, ...)
}
\arguments{
\item{UB}{The univariate basis function matrix constructed.}

\item{B}{The bernstein basis matrix.}

\item{Q2}{The \code{Q} matrix from QR decomposition of the constraint matrix.}

\item{K}{The energy matrix.}

\item{lam1}{The smoothing penalty parameter.}

\item{lam2}{The variable selection penalty parameter.}

\item{Z}{The parametric matrix.}

\item{Y}{Response variable.}

\item{ind.c}{The indexed for parametric coefficient for which standard error want to be calculated.}

\item{VS}{variable section is conducted or not.}

\item{...}{other arguments.}
}
\value{
\item{se_beta}{The standard error of linear coefficients.}
\item{Ve}{The estimated covariance matrix.}
}
\description{
This is an internal function of package \code{ggam}.
}
\details{
A sandwich formula is developed to find the standard error for \eqn{\beta}{\beta}.
The detailed algorithm is in the paper Wang et al. (2018).
}
\examples{
library(MASS)
library(grpreg)
library(Matrix)
library(BPST)
data("eg1pop_dat")
eg1_V1=eg1pop_dat[['V1']]
eg1_T1=eg1pop_dat[['T1']]
eg1pop_rho03=eg1pop_dat[['rho03']]
sam=eg1pop_rho03[sample(1:dim(eg1pop_rho03)[1],100),]
B0=basis(eg1_V1,eg1_T1, d=2, r=1, sam[,3:4])
B=B0$B
ind=B0$Ind.inside
Q2=B0$Q2
K=B0$K
Z=sam[ind,c(5:12)]
Y=sam[ind,'Y']
lam1=0.1
lam2=0
beta_se(UB=NULL,as.matrix(B),Q2,K,lam1,lam2,Z,Y,ind.c=1:6)
}
