% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_plbpsm.R
\name{cv.plbpsm}
\alias{cv.plbpsm}
\title{Cross-validation for plbpsm}
\usage{
cv.plbpsm(formula, data, kfold = 10, repeat.times = 1, family,
  group = NULL, ...)
}
\arguments{
\item{formula}{a PLBPSM formula. These are exactly like the formula for
a GLM except that univariable and bivariate smooth terms.}

\item{data}{A data frame or list containing the model response variable and
covariates required by the formula. By default the variables are taken from
environment(\code{formula}).}

\item{kfold}{number of folds splitted}

\item{repeat.times}{repeated times for cross validation}

\item{family}{This is a family object specifying the distribution and link to use in fitting
etc. See glm and family for more details.}

\item{group}{whether model identification is conducted for each covariates}

\item{...}{other arguments.}
}
\description{
Performs k-fold cross validation for GGAMs
}
