% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrplsfit.R
\name{ggrplsfit}
\alias{ggrplsfit}
\title{Variable selection in Partial Linear Bivariate penalized Spline fit in generalized spatial model}
\usage{
ggrplsfit(G, criterion, method, family, ind_c, VS,
  control = plbpsm.control(), MI, ...)
}
\arguments{
\item{G}{An object of the type returned by \code{plbpsm} when \code{fit=FALSE}.}

\item{criterion}{The criterion to choose the penalty parameter lambda. \code{"GCV"} to use
generalized cross validation method and \code{"CV"} for cross validation}

\item{method}{'ALASSO' or 'SCAD' to penalize the coefficients for parametric part.}

\item{family}{The family object, specifying the distribution and link to use.}

\item{ind_c}{The given index of covariates that are selected.}

\item{VS}{'\code{TRUE}' for using ALASSO/SCAD to select linear variables.}

\item{control}{A list of fit control parameters to replace defaults returned by \code{\link{plbpsm.control}}.
Any control parameters not supplied stay at their default values.}

\item{MI}{whether model identification is conducted or not.}

\item{...}{other arguments.}
}
\value{
A list of fit information.
}
\description{
This is an internal function of package \code{GgAM}. Bivariate penalized least squares problem
is solved with penalty parameter chosen by GCV or CV. Variable selection by using
adaptive LASSO or group SCAD is applied in parametric coefficients.
}
\details{
This is an internal function of package \code{GgAM}. We propose Iteratively Reweighted Least square based
algorithm to get the poilot estimation and then use it to get a a spline-backfitted local polynomial estimation.
The smoothing penalty parameter could
be chosen by \code{GCV} or \code{CV} using the routines: \code{\link{gplsfitGCV}}.
}
