% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsfitGCV_nb.R
\name{gplsfitGCV_nb}
\alias{gplsfitGCV_nb}
\title{Generalized Penalized Least Square Fit under GCV for negative binomial family}
\usage{
gplsfitGCV_nb(Y, B, Q2, P, UB, lambda, family, offset, r.theta = c(2, 8),
  fx, control, X = NULL, ind_c = 1:ncol(X),
  fixedSteps = (control$maxstep + 1), ...)
}
\arguments{
\item{Y}{Response variable.}

\item{B}{The bernstein basis matrix.}

\item{Q2}{The \code{Q2} matrix from QR decomposition of the transpose of the constraint matrix.}

\item{P}{The penalty matrix.}

\item{UB}{The univariate basis function matrix constructed.}

\item{lambda}{The smoothing penalty parameter.}

\item{family}{The family object, specifying the distribution and link to use.}

\item{offset}{Can be used to supply a model offset for use in fitting. Note that this offset
will always be completely ignored when predicting.}

\item{r.theta}{All the theta values given.}

\item{fx}{indicates whether the term is a fixed d.f. regression
spline (\code{TRUE}) or a penalized regression spline (\code{FALSE}).}

\item{control}{A list of fit control parameters to replace defaults returned by \code{\link{plbpsm.control}}.
Any control parameters not supplied stay at their default values.}

\item{X}{The parametric model matrix. set to '\code{NULL}' if it is not provided.}

\item{ind_c}{The vector of index to indicate the parametric part.}

\item{fixedSteps}{How many steps to take: useful when only using this routine to get rough starting
values for other methods.}

\item{...}{other arguments passed onto \code{\link{gplsfitGCV}}.}
}
\value{
A list of fit information.
}
\description{
This is an internal function of package \code{ggam}.
}
\details{
In this function, the estimator of \eqn{\theta}{\theta} is chosen to ensure that
the Pearson estimate of the scale parameter is as close as possible to 1. The other parts follow from the
routine of \code{\link{gplsfitGCV}}.
}
