% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grplsfit.R
\name{grplsfit}
\alias{grplsfit}
\title{Variable selection with Bivariate penalized Spline (GCV/CV) fit}
\usage{
grplsfit(G, criterion, method, family, ind_c, VS, MI, ...)
}
\arguments{
\item{G}{An object of the type returned by \code{plbpsm} when \code{fit=FALSE}.}

\item{criterion}{The criterion to choose the penalty parameter lambda. \code{"GCV"} to use
generalized cross validation method and \code{"CV"} for cross validation}

\item{method}{'ALASSO' or 'SCAD' to penalize the coefficients for parametric part.}

\item{family}{The family object, specifying the distribution and link to use.}

\item{ind_c}{The given index of covariates that are selected.}

\item{VS}{'\code{TRUE}' for using ALASSO/SCAD to select linear variables.}

\item{MI}{Whether model identification is conducted or not.}

\item{...}{other arguments.}
}
\value{
A list of fit information.
}
\description{
This is an internal function of package \code{ggam}. Bivariate penalized least squares problem
is solved with penalty parameter chosen by GCV or CV. Variable selection by using
adaptive LASSO or group SCAD is applied in parametric coefficients.
}
\details{
This is an internal function of package \code{ggam}. We propose a coordinate descent based
algorithm to perform the variable selection efficiently.
The smoothing penalty parameter could
be chosen by \code{GCV} or \code{CV} using the routines: \code{\link{plsfitGCV}} and
\code{\link{plsfitCV}}. In this function, the user can also choose whether to do variable selection
or not.
}
