% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plbpsm_control.R
\name{plbpsm.control}
\alias{plbpsm.control}
\title{Setting PLBPSM fitting defaults}
\usage{
plbpsm.control(delta1 = 1, delta2 = 1, trace = FALSE, maxstep = 10,
  epsilon = 1e-07)
}
\arguments{
\item{delta1}{The convergence criterion in \code{\link{gplsfitGCV}}.}

\item{delta2}{The convergence criterion in \code{\link{gplsfitGCV}}.}

\item{trace}{Set this to \code{TRUE} to turn on diagnostic output.}

\item{maxstep}{Maximum number of iterations to perform.}

\item{epsilon}{This is used for judging conversion of the loop in \code{\link{gplsfitGCV}}.}
}
\description{
This is an internal function of package \code{ggam} which allows control of the numerical options for
fitting a PLBPSM. Typically users will want to modify the defaults if model fitting fails to converge, or
if the warnings are generated which suggest a loss of numerical stability during fitting.
}
