% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{u}
\alias{u}
\title{Defining smooths in PLBPSM formulae}
\usage{
u(..., N = 2, q = 3, KnotsLocation = "quantile", knots = NULL,
  N_MI = 4, fx = FALSE, id = NULL)
}
\arguments{
\item{...}{a list of variables that are the covariates that this
smooth is a function of.}

\item{N}{Number of interior knots in generating spline matrix.}

\item{q}{Degree of polynomial spline. Default is 3.}

\item{KnotsLocation}{A character string naming the way for knots
locations. Default is "quantile". The only alternative is
"uniform".}

\item{knots}{An optional vector specifying the knots to be used in
constructing spline bases.}

\item{N_MI}{Number of interior knots in generating spline matrix in the model identification process.}

\item{fx}{indicates whether the term is a fixed d.f. regression
spline (\code{TRUE}) or a penalized regression spline (\code{FALSE}).}

\item{id}{An identifying label or number for the smooth, linking it to other smooths. Defaults to \code{NULL} for no linkage.}
}
\value{
These \code{smooth.spec} objects define uivariate smooths and are turned into bases and penalties by
\code{BasisCon} functions.
 The returned object contains the following items:
\item{term}{An array of text strings giving the names of the covariates that the term is a function of.}
\item{N}{Number of interior knots in generating spline matrix.}
\item{q}{Degree of polynomial spline. Default is 3.}
\item{knotsLocation}{A character string naming the way for knots
		locations. Default is "quantile". The only alternative is
	"uniform".}
\item{knots}{An optional vector specifying the knots to be used in
		constructing spline bases.}
\item{N_MI}{Number of interior knots in generating spline matrix in the model identification process.}
\item{fx}{\code{TRUE} if the term is to be treated as a pure regression
spline (with fixed degrees of freedom); FALSE if it is to be treated as a penalized regression spline}
\item{dim}{The dimension of the smoother - i.e. the number of covariates that it is a function of.}
\item{label}{A suitable text label for this smooth term.}
\item{id}{An identifying label or number for the smooth, linking it to other smooths. Defaults to \code{NULL} for no linkage. }
}
\description{
Function used in definition of smooth terms within \code{plbpsm} model formulae. The function does not
evaluate a (spline) smooth - it exists purely to help set up a model using spline based smooths.
}
\examples{
library(GgAM)
data(eg1pop_dat)
V=eg1pop_dat[['V2']]
Tr=eg1pop_dat[['T2']]
res <- u(x1,x2,N=2,q=3)
}
